+++ src/rs_game.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "rs_base.h"
 
-i32 thread_game(void* data);
+unsigned long thread_game(void* data);
